refactor(terminal): extract VT100 parser into terminal module (Phase T1)
Extract monolithic terminal_pty.rs into focused modules:
- Create src/panels/terminal/ module with vt100_parser.rs and mod.rs
- Extract VtPerformer (606-line csi_dispatch) to vt100_parser.rs (702 LOC)
- Extract shared types (Cell, CellStyle, MouseTrackingMode, TerminalScreen)
  and ANSI color conversion functions to terminal/mod.rs (414 LOC)
- Reduce terminal_pty.rs from 2,320 to 1,247 LOC (-1,073 LOC, -46.3%)

This Phase T1 decomposition improves terminal code organization by separating
VT100/ANSI escape sequence parsing from PTY management, making the codebase
more maintainable and easier to understand.

Total: 2,363 LOC across 3 files (+43 LOC for module structure)
