+++ src/panels/terminal_pty.rs
@@ -17,7 +17,9 @@ use ratatui::{
 use std::io::{Read, Write};
 use std::sync::{Arc, Mutex};
 use std::thread;
-use vte::{Params, Parser, Perform};
+use vte::Parser;
+
+use super::terminal::{Cell, CellStyle, MouseTrackingMode, TerminalScreen, VtPerformer};
 
 use super::Panel;
 use crate::panels::file_manager::DiskSpaceInfo;
@@ -28,15 +30,6 @@ pub struct TerminalInfo {
     pub user_host: String,                 // user@host
     pub cwd: String,                       // current directory
     pub disk_space: Option<DiskSpaceInfo>, // disk information
-}
-
-/// Mouse tracking mode
-#[derive(Clone, Copy, PartialEq)]
-enum MouseTrackingMode {
-    None,
-    Normal,      // ?1000 - clicks only
-    ButtonEvent, // ?1002 - clicks + drag
-    AnyEvent,    // ?1003 - all movements
 }
 
 /// Full-featured terminal with PTY
@@ -59,1074 +52,6 @@ pub struct Terminal {
     terminal_title: String,
     /// Initial working directory (set when terminal was created)
     initial_cwd: std::path::PathBuf,
-}
-
-/// Terminal screen state
-#[derive(Clone)]
-struct TerminalScreen {
-    /// Main line buffer - VecDeque for O(1) scroll operations
-    lines: std::collections::VecDeque<Vec<Cell>>,
-    /// Alternate screen buffer (for TUI applications)
-    alt_lines: std::collections::VecDeque<Vec<Cell>>,
-    /// Alternate screen usage flag
-    use_alt_screen: bool,
-    /// Cursor position (row, col)
-    cursor: (usize, usize),
-    /// Saved cursor position
-    saved_cursor: Option<(usize, usize)>,
-    /// Cursor visibility
-    cursor_visible: bool,
-    /// Screen dimensions
-    rows: usize,
-    cols: usize,
-    /// Current style
-    current_style: CellStyle,
-    /// Insert mode
-    #[allow(dead_code)]
-    insert_mode: bool,
-    /// Application Cursor Keys Mode (DECCKM)
-    application_cursor_keys: bool,
-    /// Mouse tracking mode
-    mouse_tracking: MouseTrackingMode,
-    /// SGR extended mouse mode (?1006)
-    sgr_mouse_mode: bool,
-    /// Bracketed paste mode (?2004)
-    bracketed_paste_mode: bool,
-    /// Text selection start (row, col)
-    selection_start: Option<(usize, usize)>,
-    /// Text selection end (row, col)
-    selection_end: Option<(usize, usize)>,
-    /// History buffer (scrollback) - VecDeque for O(1) push/pop at both ends
-    scrollback: std::collections::VecDeque<Vec<Cell>>,
-    /// View offset (0 = current screen, >0 = viewing history)
-    scroll_offset: usize,
-    /// Maximum history size (number of lines)
-    max_scrollback: usize,
-    /// Deferred line wrap
-    wrap_pending: bool,
-}
-
-/// Screen cell
-#[derive(Clone, Debug, Copy)]
-struct Cell {
-    ch: char,
-    style: CellStyle,
-}
-
-/// Cell style
-#[derive(Clone, Debug, Copy)]
-struct CellStyle {
-    fg: Color,
-    bg: Color,
-    bold: bool,
-    italic: bool,
-    underline: bool,
-    reverse: bool,
-}
-
-impl Default for CellStyle {
-    fn default() -> Self {
-        Self {
-            fg: Color::White,
-            bg: Color::Reset, // Use theme background by default
-            bold: false,
-            italic: false,
-            underline: false,
-            reverse: false,
-        }
-    }
-}
-
-impl TerminalScreen {
-    fn new(rows: usize, cols: usize) -> Self {
-        let empty_cell = Cell {
-            ch: ' ',
-            style: CellStyle::default(),
-        };
-
-        Self {
-            lines: std::collections::VecDeque::from(vec![vec![empty_cell; cols]; rows]),
-            alt_lines: std::collections::VecDeque::from(vec![vec![empty_cell; cols]; rows]),
-            use_alt_screen: false,
-            cursor: (0, 0),
-            saved_cursor: None,
-            cursor_visible: true,
-            rows,
-            cols,
-            current_style: CellStyle::default(),
-            insert_mode: false,
-            application_cursor_keys: false,
-            mouse_tracking: MouseTrackingMode::None,
-            sgr_mouse_mode: false,
-            bracketed_paste_mode: false,
-            selection_start: None,
-            selection_end: None,
-            scrollback: std::collections::VecDeque::new(),
-            scroll_offset: 0,
-            max_scrollback: 10000,
-            wrap_pending: false,
-        }
-    }
-
-    /// Get mutable reference to active buffer
-    fn active_buffer_mut(&mut self) -> &mut std::collections::VecDeque<Vec<Cell>> {
-        if self.use_alt_screen {
-            &mut self.alt_lines
-        } else {
-            &mut self.lines
-        }
-    }
-
-    /// Get reference to active buffer
-    fn active_buffer(&self) -> &std::collections::VecDeque<Vec<Cell>> {
-        if self.use_alt_screen {
-            &self.alt_lines
-        } else {
-            &self.lines
-        }
-    }
-
-    /// Switch to alternate screen
-    fn switch_to_alt_screen(&mut self) {
-        if !self.use_alt_screen {
-            self.use_alt_screen = true;
-            self.wrap_pending = false;
-            // Clear alt buffer
-            let empty_cell = Cell {
-                ch: ' ',
-                style: CellStyle::default(),
-            };
-            self.alt_lines =
-                std::collections::VecDeque::from(vec![vec![empty_cell; self.cols]; self.rows]);
-            self.cursor = (0, 0);
-        }
-    }
-
-    /// Return to main screen
-    fn switch_to_main_screen(&mut self) {
-        if self.use_alt_screen {
-            self.use_alt_screen = false;
-            self.wrap_pending = false;
-        }
-    }
-
-    /// Write character at current cursor position
-    fn put_char(&mut self, ch: char) {
-        // If there was a deferred wrap - execute it now
-        if self.wrap_pending {
-            self.wrap_pending = false;
-            self.cursor.1 = 0;
-            if self.cursor.0 + 1 >= self.rows {
-                self.scroll_up();
-            } else {
-                self.cursor.0 += 1;
-            }
-        }
-
-        let (row, col) = self.cursor;
-        let cols = self.cols;
-        let rows = self.rows;
-        let style = self.current_style;
-
-        if row < rows && col < cols {
-            let buffer = self.active_buffer_mut();
-            buffer[row][col] = Cell { ch, style };
-            // Move cursor right
-            if col + 1 >= cols {
-                // Reached last column - defer wrap
-                self.wrap_pending = true;
-            } else {
-                self.cursor.1 = col + 1;
-            }
-        }
-    }
-
-    /// Newline
-    fn newline(&mut self) {
-        self.wrap_pending = false;
-        self.cursor.1 = 0;
-        if self.cursor.0 < self.rows - 1 {
-            self.cursor.0 += 1;
-        } else {
-            // Scroll up
-            self.scroll_up();
-        }
-    }
-
-    /// Carriage return
-    fn carriage_return(&mut self) {
-        self.wrap_pending = false;
-        self.cursor.1 = 0;
-    }
-
-    /// Scroll screen up one line
-    fn scroll_up(&mut self) {
-        let cols = self.cols;
-
-        // For main buffer, save line to scrollback
-        if !self.use_alt_screen {
-            let top_line = self.lines[0].clone();
-            self.scrollback.push_back(top_line);
-
-            // Limit scrollback size - O(1) with VecDeque instead of O(n) with Vec::remove(0)
-            if self.scrollback.len() > self.max_scrollback {
-                self.scrollback.pop_front();
-            }
-        }
-
-        let buffer = self.active_buffer_mut();
-        buffer.pop_front(); // O(1) with VecDeque instead of O(n) with Vec::remove(0)
-        let empty_cell = Cell {
-            ch: ' ',
-            style: CellStyle::default(),
-        };
-        buffer.push_back(vec![empty_cell; cols]);
-    }
-
-    /// Scroll view up (into history)
-    fn scroll_view_up(&mut self, lines: usize) {
-        let max_offset = self.scrollback.len();
-        self.scroll_offset = (self.scroll_offset + lines).min(max_offset);
-    }
-
-    /// Scroll view down (to current)
-    fn scroll_view_down(&mut self, lines: usize) {
-        self.scroll_offset = self.scroll_offset.saturating_sub(lines);
-    }
-
-    /// Reset scroll to current screen
-    fn reset_scroll(&mut self) {
-        self.scroll_offset = 0;
-    }
-
-    /// Check if cell (row, col) is in current selection
-    fn is_in_selection(&self, row: usize, col: usize) -> bool {
-        let (start, end) = match (self.selection_start, self.selection_end) {
-            (Some(s), Some(e)) => (s, e),
-            _ => return false,
-        };
-
-        // Normalize: start should be before end
-        let (start, end) = if start <= end {
-            (start, end)
-        } else {
-            (end, start)
-        };
-
-        // Simple rectangular selection by lines
-        // More correct: linear selection like in regular terminals
-        if row < start.0 || row > end.0 {
-            return false;
-        }
-
-        if row == start.0 && row == end.0 {
-            // Single line
-            col >= start.1 && col <= end.1
-        } else if row == start.0 {
-            // First line
-            col >= start.1
-        } else if row == end.0 {
-            // Last line
-            col <= end.1
-        } else {
-            // Middle lines - all selected
-            true
-        }
-    }
-
-    /// Clear screen (doesn't move cursor)
-    #[allow(dead_code)]
-    fn clear_screen(&mut self) {
-        let rows = self.rows;
-        let cols = self.cols;
-        let empty_cell = Cell {
-            ch: ' ',
-            style: CellStyle::default(),
-        };
-        let buffer = self.active_buffer_mut();
-        *buffer = std::collections::VecDeque::from(vec![vec![empty_cell; cols]; rows]);
-        // Cursor stays in place (standard ED 2 behavior)
-    }
-
-    /// Move cursor
-    fn move_cursor(&mut self, row: usize, col: usize) {
-        self.wrap_pending = false;
-        self.cursor.0 = row.min(self.rows - 1);
-        self.cursor.1 = col.min(self.cols - 1);
-    }
-
-    /// Backspace
-    fn backspace(&mut self) {
-        self.wrap_pending = false;
-        if self.cursor.1 > 0 {
-            self.cursor.1 -= 1;
-        }
-    }
-
-    /// Tab
-    fn tab(&mut self) {
-        // Move cursor to next position divisible by 8
-        let next_tab = ((self.cursor.1 / 8) + 1) * 8;
-        self.cursor.1 = next_tab.min(self.cols - 1);
-    }
-
-    /// Save cursor position
-    fn save_cursor(&mut self) {
-        self.saved_cursor = Some(self.cursor);
-    }
-
-    /// Restore cursor position
-    fn restore_cursor(&mut self) {
-        if let Some(saved) = self.saved_cursor {
-            self.cursor = saved;
-            self.wrap_pending = false;
-        }
-    }
-}
-
-/// VT Performer for handling escape sequences
-struct VtPerformer {
-    screen: Arc<Mutex<TerminalScreen>>,
-    pending_backslash: bool,
-}
-
-impl Perform for VtPerformer {
-    fn print(&mut self, ch: char) {
-        // Filter control characters that shouldn't be displayed
-        // (except printable characters)
-        if ch.is_control() && ch != '\t' && ch != '\n' && ch != '\r' {
-            return;
-        }
-
-        // Handle bash readline markers \[ and \]
-        if self.pending_backslash {
-            self.pending_backslash = false;
-            // If backslash is followed by [ or ], skip both characters
-            if ch == '[' || ch == ']' {
-                return;
-            }
-            // Otherwise print deferred backslash and current character
-            if let Ok(mut screen) = self.screen.lock() {
-                screen.put_char('\\');
-                screen.put_char(ch);
-            }
-            return;
-        }
-
-        // If we encounter backslash, defer it
-        if ch == '\\' {
-            self.pending_backslash = true;
-            return;
-        }
-
-        if let Ok(mut screen) = self.screen.lock() {
-            screen.put_char(ch);
-        }
-    }
-
-    fn execute(&mut self, byte: u8) {
-        if let Ok(mut screen) = self.screen.lock() {
-            match byte {
-                b'\n' => screen.newline(),
-                b'\r' => screen.carriage_return(),
-                b'\x08' => screen.backspace(),
-                b'\t' => screen.tab(),
-                b'\x07' => {
-                    // Bell character - forward to parent terminal
-                    print!("\x07");
-                    let _ = std::io::stdout().flush();
-                }
-                _ => {}
-            }
-        }
-    }
-
-    fn hook(&mut self, _params: &Params, _intermediates: &[u8], _ignore: bool, _c: char) {}
-
-    fn put(&mut self, _byte: u8) {}
-
-    fn unhook(&mut self) {}
-
-    fn osc_dispatch(&mut self, _params: &[&[u8]], _bell_terminated: bool) {}
-
-    fn csi_dispatch(&mut self, params: &Params, intermediates: &[u8], _ignore: bool, c: char) {
-        // Handle private sequences (start with '?')
-        if !intermediates.is_empty() && intermediates[0] == b'?' {
-            if let Ok(mut screen) = self.screen.lock() {
-                // Get private sequence number
-                let mode = params
-                    .iter()
-                    .next()
-                    .and_then(|p| p.first())
-                    .copied()
-                    .unwrap_or(0);
-
-                match (mode, c) {
-                    (1049, 'h') => {
-                        // Switch to alternate screen and save cursor
-                        screen.saved_cursor = Some(screen.cursor);
-                        screen.switch_to_alt_screen();
-                    }
-                    (1049, 'l') => {
-                        // Return to main screen and restore cursor
-                        screen.switch_to_main_screen();
-                        if let Some(saved) = screen.saved_cursor {
-                            screen.cursor = saved;
-                            screen.saved_cursor = None;
-                        }
-                    }
-                    (47, 'h') => {
-                        // Switch to alternate screen (without saving cursor)
-                        screen.switch_to_alt_screen();
-                    }
-                    (47, 'l') => {
-                        // Return to main screen
-                        screen.switch_to_main_screen();
-                    }
-                    (25, 'h') => {
-                        // Show cursor
-                        screen.cursor_visible = true;
-                    }
-                    (25, 'l') => {
-                        // Hide cursor
-                        screen.cursor_visible = false;
-                    }
-                    (1, 'h') => {
-                        // DECCKM - Application Cursor Keys Mode ON
-                        screen.application_cursor_keys = true;
-                    }
-                    (1, 'l') => {
-                        // DECCKM - Application Cursor Keys Mode OFF
-                        screen.application_cursor_keys = false;
-                    }
-                    // Mouse tracking modes
-                    (1000, 'h') => {
-                        // Normal tracking mode ON
-                        screen.mouse_tracking = MouseTrackingMode::Normal;
-                    }
-                    (1000, 'l') => {
-                        // Normal tracking mode OFF
-                        screen.mouse_tracking = MouseTrackingMode::None;
-                    }
-                    (1002, 'h') => {
-                        // Button event tracking mode ON
-                        screen.mouse_tracking = MouseTrackingMode::ButtonEvent;
-                    }
-                    (1002, 'l') => {
-                        // Button event tracking mode OFF
-                        screen.mouse_tracking = MouseTrackingMode::None;
-                    }
-                    (1003, 'h') => {
-                        // Any event tracking mode ON
-                        screen.mouse_tracking = MouseTrackingMode::AnyEvent;
-                    }
-                    (1003, 'l') => {
-                        // Any event tracking mode OFF
-                        screen.mouse_tracking = MouseTrackingMode::None;
-                    }
-                    (1006, 'h') => {
-                        // SGR extended mouse mode ON
-                        screen.sgr_mouse_mode = true;
-                    }
-                    (1006, 'l') => {
-                        // SGR extended mouse mode OFF
-                        screen.sgr_mouse_mode = false;
-                    }
-                    (2004, 'h') => {
-                        // Bracketed paste mode ON
-                        screen.bracketed_paste_mode = true;
-                    }
-                    (2004, 'l') => {
-                        // Bracketed paste mode OFF
-                        screen.bracketed_paste_mode = false;
-                    }
-                    _ => {
-                        // Ignore other private sequences
-                    }
-                }
-            }
-            return;
-        }
-
-        // Ignore other intermediate bytes
-        if !intermediates.is_empty() {
-            return;
-        }
-
-        if let Ok(mut screen) = self.screen.lock() {
-            match c {
-                'H' | 'f' => {
-                    // Move cursor
-                    let row = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    let col = params
-                        .iter()
-                        .nth(1)
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.move_cursor(row.saturating_sub(1), col.saturating_sub(1));
-                }
-                'J' => {
-                    // ED - Erase in Display
-                    let param = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(0);
-                    let (row, col) = screen.cursor;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    match param {
-                        0 => {
-                            // Clear from cursor to end of screen
-                            let buffer = screen.active_buffer_mut();
-                            let buf_rows = buffer.len();
-
-                            // Clear rest of current line
-                            if row < buf_rows {
-                                let buf_cols = buffer[row].len();
-                                for i in col..buf_cols {
-                                    buffer[row][i] = empty_cell;
-                                }
-                            }
-                            // Clear all lines below
-                            for r in (row + 1)..buf_rows {
-                                let buf_cols = buffer[r].len();
-                                for c in 0..buf_cols {
-                                    buffer[r][c] = empty_cell;
-                                }
-                            }
-                        }
-                        1 => {
-                            // Clear from start of screen to cursor
-                            let buffer = screen.active_buffer_mut();
-                            let buf_rows = buffer.len();
-
-                            // Clear all lines above
-                            for r in 0..row.min(buf_rows) {
-                                let buf_cols = buffer[r].len();
-                                for c in 0..buf_cols {
-                                    buffer[r][c] = empty_cell;
-                                }
-                            }
-                            // Clear current line up to and including cursor
-                            if row < buf_rows {
-                                let buf_cols = buffer[row].len();
-                                for i in 0..=col.min(buf_cols.saturating_sub(1)) {
-                                    buffer[row][i] = empty_cell;
-                                }
-                            }
-                        }
-                        2 => {
-                            // Clear entire screen and move cursor to (0,0)
-                            let buffer = screen.active_buffer_mut();
-                            for row in buffer.iter_mut() {
-                                row.fill(empty_cell);
-                            }
-                            // Move cursor to home position (compatibility with old behavior)
-                            screen.cursor = (0, 0);
-                        }
-                        3 => {
-                            // Clear entire screen and scrollback
-                            let is_alt = screen.use_alt_screen;
-                            let buffer = screen.active_buffer_mut();
-                            for row in buffer.iter_mut() {
-                                row.fill(empty_cell);
-                            }
-                            // Clear scrollback only for main screen
-                            if !is_alt {
-                                screen.scrollback.clear();
-                            }
-                            screen.cursor = (0, 0);
-                        }
-                        _ => {}
-                    }
-                }
-                'K' => {
-                    // EL - Erase in Line
-                    let param = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(0);
-                    let (row, col) = screen.cursor;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    let buffer = screen.active_buffer_mut();
-                    if row < buffer.len() {
-                        let buf_cols = buffer[row].len();
-                        match param {
-                            0 => {
-                                // From cursor to end of line
-                                for i in col..buf_cols {
-                                    buffer[row][i] = empty_cell;
-                                }
-                            }
-                            1 => {
-                                // From start of line to cursor (inclusive)
-                                for i in 0..=col.min(buf_cols.saturating_sub(1)) {
-                                    buffer[row][i] = empty_cell;
-                                }
-                            }
-                            2 => {
-                                // Entire line
-                                for i in 0..buf_cols {
-                                    buffer[row][i] = empty_cell;
-                                }
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-                'P' => {
-                    // DCH - Delete Character
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    let (row, col) = screen.cursor;
-                    let cols = screen.cols;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    let buffer = screen.active_buffer_mut();
-                    // Shift characters left from deleted position
-                    for i in col..(cols - n) {
-                        buffer[row][i] = buffer[row][i + n];
-                    }
-
-                    // Fill freed space with blanks
-                    for i in (cols - n)..cols {
-                        buffer[row][i] = empty_cell;
-                    }
-                }
-                'X' => {
-                    // ECH - Erase Character
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    let (row, col) = screen.cursor;
-                    let cols = screen.cols;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    let buffer = screen.active_buffer_mut();
-                    for i in col..(col + n).min(cols) {
-                        buffer[row][i] = empty_cell;
-                    }
-                }
-                '@' => {
-                    // ICH - Insert Character (shift characters right)
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    let (row, col) = screen.cursor;
-                    let cols = screen.cols;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    let buffer = screen.active_buffer_mut();
-                    // Shift characters right
-                    if col + n < cols {
-                        for i in (col + n..cols).rev() {
-                            buffer[row][i] = buffer[row][i - n];
-                        }
-                    }
-
-                    // Insert blanks at freed positions
-                    for i in col..(col + n).min(cols) {
-                        buffer[row][i] = empty_cell;
-                    }
-                }
-                'L' => {
-                    // IL - Insert Lines (insert blank lines)
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    let row = screen.cursor.0;
-                    let cols = screen.cols;
-                    let rows = screen.rows;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    let buffer = screen.active_buffer_mut();
-                    if row < buffer.len() {
-                        // Delete n lines from bottom
-                        for _ in 0..n.min(rows - row) {
-                            if buffer.len() > row {
-                                buffer.pop_back();
-                            }
-                        }
-                        // Insert n blank lines at cursor position
-                        for _ in 0..n.min(rows - row) {
-                            buffer.insert(row, vec![empty_cell; cols]);
-                        }
-                    }
-                }
-                'M' => {
-                    // DL - Delete Lines (delete lines)
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    let row = screen.cursor.0;
-                    let cols = screen.cols;
-                    let rows = screen.rows;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    let buffer = screen.active_buffer_mut();
-                    if row < buffer.len() {
-                        // Delete n lines at cursor position
-                        for _ in 0..n.min(buffer.len() - row) {
-                            if row < buffer.len() {
-                                buffer.remove(row);
-                            }
-                        }
-                        // Add n blank lines at bottom
-                        while buffer.len() < rows {
-                            buffer.push_back(vec![empty_cell; cols]);
-                        }
-                    }
-                }
-                'S' => {
-                    // SU - Scroll Up (scroll screen up)
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    let cols = screen.cols;
-                    let rows = screen.rows;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    let buffer = screen.active_buffer_mut();
-                    for _ in 0..n.min(rows) {
-                        if !buffer.is_empty() {
-                            buffer.pop_front(); // O(1) with VecDeque
-                        }
-                        buffer.push_back(vec![empty_cell; cols]);
-                    }
-                }
-                'T' => {
-                    // SD - Scroll Down (scroll screen down)
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    let cols = screen.cols;
-                    let rows = screen.rows;
-                    let empty_cell = Cell {
-                        ch: ' ',
-                        style: screen.current_style,
-                    };
-
-                    let buffer = screen.active_buffer_mut();
-                    for _ in 0..n.min(rows) {
-                        if buffer.len() >= rows {
-                            buffer.pop_back();
-                        }
-                        buffer.push_front(vec![empty_cell; cols]); // O(1) with VecDeque
-                    }
-                }
-                'A' => {
-                    // Cursor up
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.wrap_pending = false;
-                    screen.cursor.0 = screen.cursor.0.saturating_sub(n);
-                }
-                'B' => {
-                    // Cursor down
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.wrap_pending = false;
-                    screen.cursor.0 = (screen.cursor.0 + n).min(screen.rows - 1);
-                }
-                'C' => {
-                    // Cursor right
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.wrap_pending = false;
-                    screen.cursor.1 = (screen.cursor.1 + n).min(screen.cols - 1);
-                }
-                'D' => {
-                    // Cursor left
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.wrap_pending = false;
-                    screen.cursor.1 = screen.cursor.1.saturating_sub(n);
-                }
-                'E' => {
-                    // CNL - Cursor Next Line
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.wrap_pending = false;
-                    screen.cursor.0 = (screen.cursor.0 + n).min(screen.rows - 1);
-                    screen.cursor.1 = 0;
-                }
-                'F' => {
-                    // CPL - Cursor Previous Line
-                    let n = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.wrap_pending = false;
-                    screen.cursor.0 = screen.cursor.0.saturating_sub(n);
-                    screen.cursor.1 = 0;
-                }
-                'G' => {
-                    // CHA - Cursor Horizontal Absolute
-                    let col = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.wrap_pending = false;
-                    screen.cursor.1 = col.saturating_sub(1).min(screen.cols - 1);
-                }
-                'd' => {
-                    // VPA - Vertical Position Absolute
-                    let row = params
-                        .iter()
-                        .next()
-                        .and_then(|p| p.first())
-                        .copied()
-                        .unwrap_or(1) as usize;
-                    screen.wrap_pending = false;
-                    screen.cursor.0 = row.saturating_sub(1).min(screen.rows - 1);
-                }
-                'm' => {
-                    // SGR - set style
-                    // Collect all parameters into one vector to handle 38;5;N and 48;5;N
-                    let all_params: Vec<u16> =
-                        params.iter().flat_map(|p| p.iter().copied()).collect();
-                    let mut i = 0;
-                    while i < all_params.len() {
-                        let p = all_params[i];
-                        match p {
-                            0 => screen.current_style = CellStyle::default(),
-                            1 => screen.current_style.bold = true,
-                            3 => screen.current_style.italic = true,
-                            4 => screen.current_style.underline = true,
-                            7 => screen.current_style.reverse = true,
-                            22 => screen.current_style.bold = false,
-                            23 => screen.current_style.italic = false,
-                            24 => screen.current_style.underline = false,
-                            27 => screen.current_style.reverse = false,
-                            30..=37 => {
-                                screen.current_style.fg = ansi_to_color(p - 30);
-                            }
-                            38 => {
-                                // 256-color or RGB foreground
-                                if i + 2 < all_params.len() && all_params[i + 1] == 5 {
-                                    // 38;5;N - 256-color
-                                    let color_idx = all_params[i + 2];
-                                    screen.current_style.fg = ansi_256_to_color(color_idx);
-                                    i += 2;
-                                } else if i + 4 < all_params.len() && all_params[i + 1] == 2 {
-                                    // 38;2;R;G;B - True Color (24-bit)
-                                    let r = all_params[i + 2] as u8;
-                                    let g = all_params[i + 3] as u8;
-                                    let b = all_params[i + 4] as u8;
-                                    screen.current_style.fg = Color::Rgb(r, g, b);
-                                    i += 4;
-                                }
-                            }
-                            39 => {
-                                // Reset foreground to default
-                                screen.current_style.fg = Color::Reset;
-                            }
-                            40..=47 => {
-                                screen.current_style.bg = ansi_to_color(p - 40);
-                            }
-                            48 => {
-                                // 256-color or RGB background
-                                if i + 2 < all_params.len() && all_params[i + 1] == 5 {
-                                    // 48;5;N - 256-color
-                                    let color_idx = all_params[i + 2];
-                                    screen.current_style.bg = ansi_256_to_color(color_idx);
-                                    i += 2;
-                                } else if i + 4 < all_params.len() && all_params[i + 1] == 2 {
-                                    // 48;2;R;G;B - True Color (24-bit)
-                                    let r = all_params[i + 2] as u8;
-                                    let g = all_params[i + 3] as u8;
-                                    let b = all_params[i + 4] as u8;
-                                    screen.current_style.bg = Color::Rgb(r, g, b);
-                                    i += 4;
-                                }
-                            }
-                            49 => {
-                                // Reset background to default
-                                screen.current_style.bg = Color::Reset;
-                            }
-                            90..=97 => {
-                                screen.current_style.fg = ansi_to_bright_color(p - 90);
-                            }
-                            100..=107 => {
-                                screen.current_style.bg = ansi_to_bright_color(p - 100);
-                            }
-                            _ => {}
-                        }
-                        i += 1;
-                    }
-                }
-                's' => {
-                    // Save cursor position
-                    screen.save_cursor();
-                }
-                'u' => {
-                    // Restore cursor position
-                    screen.restore_cursor();
-                }
-                'r' => {
-                    // DECSTBM - Set scrolling region (ignore but don't break)
-                }
-                'l' | 'h' => {
-                    // Set/Reset Mode (ignore but don't break)
-                }
-                _ => {}
-            }
-        }
-    }
-
-    fn esc_dispatch(&mut self, _intermediates: &[u8], _ignore: bool, _byte: u8) {}
-}
-
-/// Convert ANSI color to ratatui Color
-fn ansi_to_color(code: u16) -> Color {
-    match code {
-        0 => Color::Black,
-        1 => Color::Red,
-        2 => Color::Green,
-        3 => Color::Yellow,
-        4 => Color::Blue,
-        5 => Color::Magenta,
-        6 => Color::Cyan,
-        7 => Color::White,
-        _ => Color::White,
-    }
-}
-
-/// Convert bright ANSI color to ratatui Color
-fn ansi_to_bright_color(code: u16) -> Color {
-    match code {
-        0 => Color::DarkGray,
-        1 => Color::LightRed,
-        2 => Color::LightGreen,
-        3 => Color::LightYellow,
-        4 => Color::LightBlue,
-        5 => Color::LightMagenta,
-        6 => Color::LightCyan,
-        7 => Color::White,
-        _ => Color::White,
-    }
-}
-
-/// Convert 256-color index to ratatui Color
-fn ansi_256_to_color(code: u16) -> Color {
-    match code {
-        // Basic 16 colors (0-15)
-        0 => Color::Black,
-        1 => Color::Red,
-        2 => Color::Green,
-        3 => Color::Yellow,
-        4 => Color::Blue,
-        5 => Color::Magenta,
-        6 => Color::Cyan,
-        7 => Color::White,
-        8 => Color::DarkGray,
-        9 => Color::LightRed,
-        10 => Color::LightGreen,
-        11 => Color::LightYellow,
-        12 => Color::LightBlue,
-        13 => Color::LightMagenta,
-        14 => Color::LightCyan,
-        15 => Color::White,
-        // 216 colors (6x6x6 cube) - indices 16-231
-        16..=231 => {
-            let idx = code - 16;
-            let r = (idx / 36) as u8;
-            let g = ((idx % 36) / 6) as u8;
-            let b = (idx % 6) as u8;
-            // Convert 0-5 to 0-255
-            let r = if r == 0 { 0 } else { 55 + r * 40 };
-            let g = if g == 0 { 0 } else { 55 + g * 40 };
-            let b = if b == 0 { 0 } else { 55 + b * 40 };
-            Color::Rgb(r, g, b)
-        }
-        // 24 grayscale gradations (232-255)
-        232..=255 => {
-            let gray = 8 + (code - 232) as u8 * 10;
-            Color::Rgb(gray, gray, gray)
-        }
-        _ => Color::White,
-    }
 }
 
 impl Terminal {
